refactor!: add tls, rgpc server file
combine ssl.cnf to 1, update makefile
update readme
fix tests don't run
