+++ cmd/main.go
@@ -1,72 +1,14 @@
 package main
 
 import (
-	"crypto/tls"
-	"crypto/x509"
 	"flag"
 	"fmt"
 	"log"
 	"net"
-	"os"
 
-	"dungtl2003/snowflake-uuid/internal/data"
-	pb "dungtl2003/snowflake-uuid/internal/pb"
 	"dungtl2003/snowflake-uuid/internal/server"
 	"dungtl2003/snowflake-uuid/internal/tools"
-
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
 )
-
-const (
-	serverCertFile   = "x509/server_cert.pem"
-	serverKeyFile    = "x509/server_key.pem"
-	clientCACertFile = "x509/ca_cert.pem"
-)
-
-func loadTlsCredentials() (credentials.TransportCredentials, error) {
-	pemClientCA, err := os.ReadFile(data.Path(clientCACertFile))
-	if err != nil {
-		return nil, err
-	}
-
-	certPools := x509.NewCertPool()
-	if !certPools.AppendCertsFromPEM(pemClientCA) {
-		return nil, fmt.Errorf("failed to add client CA's certificate")
-	}
-
-	cert, err := tls.LoadX509KeyPair(data.Path(serverCertFile), data.Path(serverKeyFile))
-	if err != nil {
-		return nil, err
-	}
-
-	config := &tls.Config{
-		Certificates: []tls.Certificate{cert},
-		ClientAuth:   tls.RequireAndVerifyClientCert,
-		ClientCAs:    certPools,
-	}
-	return credentials.NewTLS(config), nil
-}
-
-func runGrpcServer(idGeneratorServer *server.IdGeneratorServer, enableTls bool, lis net.Listener) error {
-	var grpcServer *grpc.Server
-	if !enableTls {
-		grpcServer = grpc.NewServer()
-	} else {
-		tlsCredentials, err := loadTlsCredentials()
-		if err != nil {
-			return err
-		}
-
-		grpcServer = grpc.NewServer(
-			grpc.Creds(tlsCredentials),
-		)
-	}
-
-	pb.RegisterIdGeneratorServer(grpcServer, idGeneratorServer)
-
-	return grpcServer.Serve(lis)
-}
 
 func main() {
 	port := flag.Int("port", 0, "the server port")
@@ -87,7 +29,7 @@ func main() {
 	}
 
 	log.Println("Creating ID generator server")
-	idGeneratorServer, err := server.New(config.WorkerId, config.Epoch)
+	idGeneratorServer, err := server.NewIdGeneratorServer(config.WorkerId, config.Epoch)
 	if err != nil {
 		log.Fatalf("Failed to create ID generator server: %v", err)
 	}
@@ -99,7 +41,7 @@ func main() {
 	}
 
 	log.Printf("Serving GRPC server on port %d", *port)
-	err = runGrpcServer(idGeneratorServer, *enableTLS, lis)
+	err = server.RunGrpcServer(idGeneratorServer, *enableTLS, lis)
 	if err != nil {
 		log.Fatalf("Failed to serve grpc server on port %d: %v", *port, err)
 	}
