+++ internal/data/x509/openssl.cnf
@@ -21,7 +21,7 @@ subjectAltName = @server_alt_names
 [ server_alt_names ]
 DNS.1 = chatapp.com
 DNS.2 = *.chatapp.com
-IP.1 = 0.0.0.0
+DNS.3 = localhost
 
 [ client ]
 basicConstraints = critical,CA:FALSE
@@ -33,5 +33,4 @@ subjectAltName = @client_alt_names
 [ client_alt_names ]
 DNS.1 = chatapp.com
 DNS.2 = *.chatapp.com
-IP.1 = 0.0.0.0
-
+DNS.3 = localhost
